Extracting PlayerDetail into its own user control
